# Optimize slice appending by using slice expansion