Optimize slice appending by using slice expansion (`append(data, v...)`) instead of a loop.

The benchmark results show a minor improvement in execution time for sparse sketches.

```
goos: darwin
goarch: arm64
pkg: github.com/axiomhq/hyperloglog
cpu: Apple M3
                                                    │   old.txt    │               new.txt                │
                                                    │    sec/op    │    sec/op     vs base                │
_MarshalBinary/Precision_12/Size_100/Sparse-8         350.0n ± 19%   250.0n ± 60%  -28.57% (p=0.014 n=10)
_MarshalBinary/Precision_12/Size_100/NoSparse-8       1.534µ ±  9%   1.575µ ±  6%        ~ (p=0.541 n=10)
_MarshalBinary/Precision_12/Size_10000/Sparse-8       1.554µ ±  6%   1.542µ ±  4%        ~ (p=0.323 n=10)
_MarshalBinary/Precision_12/Size_10000/NoSparse-8     1.525µ ±  3%   1.512µ ± 14%        ~ (p=0.491 n=10)
_MarshalBinary/Precision_12/Size_1000000/Sparse-8     1.533µ ±  3%   1.508µ ±  1%   -1.63% (p=0.029 n=10)
_MarshalBinary/Precision_12/Size_1000000/NoSparse-8   1.525µ ±  2%   1.529µ ± 17%        ~ (p=0.616 n=10)
_MarshalBinary/Precision_14/Size_100/Sparse-8         579.2n ± 15%   654.1n ± 18%        ~ (p=0.122 n=10)
_MarshalBinary/Precision_14/Size_100/NoSparse-8       5.466µ ±  6%   5.484µ ±  4%        ~ (p=0.643 n=10)
_MarshalBinary/Precision_14/Size_10000/Sparse-8       5.904µ ±  8%   5.821µ ±  2%        ~ (p=0.541 n=10)
_MarshalBinary/Precision_14/Size_10000/NoSparse-8     5.980µ ±  4%   5.667µ ±  2%   -5.23% (p=0.002 n=10)
_MarshalBinary/Precision_14/Size_1000000/Sparse-8     5.863µ ±  8%   5.804µ ±  5%        ~ (p=0.362 n=10)
_MarshalBinary/Precision_14/Size_1000000/NoSparse-8   5.854µ ±  4%   5.792µ ±  7%        ~ (p=0.927 n=10)
_MarshalBinary/Precision_16/Size_100/Sparse-8         608.3n ± 58%   604.1n ± 23%        ~ (p=0.698 n=10)
_MarshalBinary/Precision_16/Size_100/NoSparse-8       21.12µ ±  3%   21.33µ ±  4%        ~ (p=0.529 n=10)
_MarshalBinary/Precision_16/Size_10000/Sparse-8       7.900µ ±  3%   1.296µ ± 65%  -83.60% (p=0.000 n=10)
_MarshalBinary/Precision_16/Size_10000/NoSparse-8     20.70µ ±  1%   20.37µ ±  2%        ~ (p=0.072 n=10)
_MarshalBinary/Precision_16/Size_1000000/Sparse-8     20.66µ ±  3%   20.38µ ±  4%   -1.33% (p=0.045 n=10)
_MarshalBinary/Precision_16/Size_1000000/NoSparse-8   20.75µ ±  2%   20.34µ ±  1%   -1.99% (p=0.015 n=10)
geomean                                               3.599µ         3.192µ        -11.30%
```